Add a 'Using Mina' page to docs
After gathering some community feedback, there was some confusion with
sending an initial transaction in the `connecting` page. This sparked
some motivation to refactor the `connecting` docs a bit and limit it's
discussion towards just connecting and syncing to the network. A new
page has been created named `Using Mina` that specifies some basic usage
of the Mina client such as importing an account, checking current
accounts, and making a payment. This decoupling makes for a better UX
for new users.